+++ __init__.py
@@ -1,13 +1,13 @@
-from . import bet
+from thread import Thread
 
 thread = Thread()
 thread.load()
 
 while True:
-    line = raw_input("> ")
+    line = input("> ")
     if line == "exit":
         break
-    
+
     if thread.interpret(line.split(" ")) != None:
         file_append("betting_history.txt", line + "\n")
         thread.save()
