+++ updates.txt
@@ -146,4 +146,3 @@ Frostfire, Kalle, MBOUltimate, TheEvoker & Tue27: +$80.77
 Threefolder: +$80.76
 Lamp, Robert, TheTurk & hPerks: +$53.84
 HiGuy & Jackrb: -$100.00
-
